requirements.txt
jinja2
json-repair
moviepy
pandas
openai
pyarrow

trustcall

langchain
langsmith 
langchain-core
langchain-openai
langgraph
langgraph-swarm
langchain-community
langchain-experimental
langgraph-reflection
langchain_google_genai
langchain_chroma
langchain_openai

matplotlib
wordcloud
squarify

streamlit
json-repair
scikit-learn

trustcall
python-dotenv

pydantic

pypdf

pubchempy
unstructured[docx]